requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
greenlet==3.2.1
idna==3.10
playwright==1.52.0
pyee==13.0.0
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0